ref: create a schema joi for each model
